+++ src/Calendar/TableRow.tsx
@@ -27,6 +27,7 @@ const TableRow: RsRefForwardingComponent<'div', TableRowProps> = React.forwardRe
       isoWeek,
       onMouseMove,
       onSelect,
+      cellClassName,
       renderCell,
       locale: overrideLocale,
       showWeekNumbers
@@ -66,6 +67,8 @@ const TableRow: RsRefForwardingComponent<'div', TableRowProps> = React.forwardRe
           ? isStartSelected || isEndSelected
           : DateUtils.isSameDay(thisDate, selected);
 
+        // TODO-Doma Move those logic that's for DatePicker/DateRangePicker to a separate component
+        //           Calendar is not supposed to be reused this way
         let inRange = false;
         // for Selected
         if (selectedStartDate && selectedEndDate) {
@@ -99,15 +102,18 @@ const TableRow: RsRefForwardingComponent<'div', TableRowProps> = React.forwardRe
           }
         }
 
-        const classes = prefix('cell', {
-          'cell-un-same-month': unSameMonth,
-          'cell-is-today': isToday,
-          'cell-selected': isSelected,
-          'cell-selected-start': isStartSelected,
-          'cell-selected-end': isEndSelected,
-          'cell-in-range': !unSameMonth && inRange,
-          'cell-disabled': disabled
-        });
+        const classes = merge(
+          prefix('cell', {
+            'cell-un-same-month': unSameMonth,
+            'cell-is-today': isToday,
+            'cell-selected': isSelected,
+            'cell-selected-start': isStartSelected,
+            'cell-selected-end': isEndSelected,
+            'cell-in-range': !unSameMonth && inRange,
+            'cell-disabled': disabled
+          }),
+          cellClassName?.(thisDate)
+        );
 
         const title = formatDate
           ? formatDate(thisDate, formatStr)